Battle UI almost done

+++ SimpleApplication/Scene.h
@@ -1,3 +1,4 @@
+// Lucas de Souza Góes (C) 2020
 #pragma once
 #include <string>
 
@@ -15,7 +16,7 @@ protected:
 	}
 	inline void DrawToScreen(wchar_t* screen, const int screenWidth, int posX, int posY, std::string content)
 	{
-		for (int i = 0; i < content.length(); i++)
+		for (size_t i = 0; i < content.length(); i++)
 		{
 			screen[posY * screenWidth + (posX+i)] = content[i];
 		}
